# documentation/swagger.yaml
swagger: "2.0"
info: 
  version: "1.0.0"
  title: GameLabz
  description: |
    ## Required software:

    - Java 1.8+
    - PostgreSQL 9+
    - Apache Maven 3+
    
    ## Build:
    CD to the pom folder
    
    ```
    mvn clean package -DskipTests
    ```
    The jar will be in the /target folder.
    
    This SQL script has to be executed once when the server is deployed for the first time:
    
    ```
    drop database if exists "gameLabz";
    create database "gameLabz";
    ```
    
    The required tables will be created by the application.
    
    ## Usage: 
    ```
    java -jar gameLabz-1.0-SNAPSHOT.jar [parameters]
    Parameters:
    --server.port: the server port, default: 8080
    --database.username: the username for the PostgreSQL database, mandatory
    --database.password: the password for the PostgreSQL database, mandatory
    ```
  contact:
    name: Razman Tamas
host: localhost:8080
basePath: /gameLabz
schemes:
  - http
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create new user
      description: A user is being registered and after that it has to subscribe to /friendRequest and /gameGift topics to receive notifications
      operationId: registerUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: userBody
          description: User that needs to be registered
          required: true
          schema:
            $ref: "#/definitions/UserForRegister"
      responses:
        201:
          description: A new user has been successfully created
          schema:
            $ref: "#/definitions/ResponseUser"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
...
  /users/{userName}:
    put:
      tags:
        - users
      summary: Updates user's password and/or email
      description: Finds user by id and updates password and/or email with the ones in the input
      operationId: updateUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the user that needs to be updated
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: userBody
          description: New password and/or email to replace the old one
          required: true
          schema:
            $ref: "#/definitions/UserForUpdate"
      responses:
        200:
          description: User has been successfully updated
          schema:
            $ref: "#/definitions/ResponseUser"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
          
    delete:
      tags:
        - users
      summary: Deletes a user
      operationId: deleteUser
      parameters:
        - in: path
          name: userName
          description: Username of the user that needs to be deleted
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: User has been successfully deleted
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
  /users/{userName}/logout:
    post:
      tags:
        - users
      summary: Logs out a user
      operationId: logoutUser
      parameters:
        - in: path
          name: userName
          description: Username of the user that needs to log out
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Logout successful
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
    
    
    
#----------------------- GAME ------------------------
...
  /share/{sharedLinkId}:
    put:
      tags:
        - userProfile
      summary: Adds a point to a user who ownes the link
      description: Every user has a personal link which they can share and they get points each time the link is accessed
      operationId: clickedOnLink
      parameters:
        - in: path
          name: sharedLinkId
          description: This ID is used to distinguish each user's personal link
          required: true
          type: string
      responses:
        200:
          description: Link has been accessed successfully
        400:
          description: Invalid link
          headers:
            error_message:
              type: string
              description: The error message
  /users/{userName}/profile:
    get:
      tags:
        - userProfile
      summary: A user views his own profile
      operationId: getOwnProfile
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to view his own profile
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Getting user's profile was successful
          schema:
            $ref: "#/definitions/UserProfile"
        400:
          description: Invalid username
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        500:
          description: Error occured while fetching profile
          headers:
            error_message:
              type: string
              description: The error message
  /users/{userName}/profile/{friendUserName}:
    get:
      tags:
        - userProfile
      summary: A user views one of his friend's profile
      operationId: getFriendProfile
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to view his friend's profile
          required: true
          type: string
        - in: path
          name: friendUserName
          description: Username of the user whose profile is being viewed
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Getting user's profile was successful
          schema:
            $ref: "#/definitions/UserProfile"
        400:
          description: Invalid username
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid or the two users are not friends
          headers:
            error_message:
              type: string
              description: The error message
        500:
          description: Error occured while fetching profile
          headers:
            error_message:
              type: string
              description: The error message
        
  


#----------------------- GAME ORDER ------------------------
...
  /users/{userName}/gameRating/{gameId}:
    post:
      tags:
        - gameRating
      summary: A user rates a game he owns
      operationId: rateGame
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to rate a game
          required: true
          type: string
        - in: path
          name: gameId
          description: Id of the game that is being rated
          required: true
          type: integer
          format: int64
        - in: query
          name: rating
          description: The value of the rating in [0, 100]
          required: true
          type: number
          format: double
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Game was successfully rated
          schema:
            $ref: "#/definitions/Rating"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid or user doesn't own the game
          headers:
            error_message:
              type: string
              description: The error message
        500:
          description: Error occured while fetching the rating
          headers:
            error_message:
              type: string
              description: The error message



    
#----------------------- DEFINITIONS ------------------------
definitions:
  ResponseUser:
      type: object
      properties:
        userName:
          type: string
        email:
          type: string
        userType:
          type: string
          enum:
            - ADMIN
            - CLIENT
  UserForRegister:
      type: object
      properties:
        userName:
          type: string
        password:
          type: string
        email:
          type: string
...
  UserForUpdate:
      type: object
      properties:
        password: 
          type: string
        email:
          type: string
  Game:
      type: object
      properties:
        name:
          type: string
        description: 
          type: string
        releaseDate:
          type: integer
          format: int64
        gameGenre:
          type: string
          enum:
            - RPG
            - FPS
            - MOBA
...
  GameOffer:
      type: object
      properties: 
        games:
          type: array
          items:
            $ref: "#/definitions/Game"
        priceCash:
          type: number
          format: double
        pricePoints:
          type: integer
          format: int64
        promotion:
          type: number
          format: double
        offerType:
          type: string
          enum:
            - BUNDLE
            - SINGLE
...
  UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        friends:
          type: array
          items:
            $ref: "#/definitions/Friends"
        ownedGames:
          type: array
          items: 
            $ref: "#/definitions/Game"
        generatedKeys:
          type: array
          items:
            $ref: "#/definitions/GeneratedKey"
        points:
          type: integer
          format: int64
        sharedLinkId:
          type: string
        sharedLink:
          type: string
  Friends:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        friend:
          $ref: "#/definitions/ResponseUser"
...
  GameOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        gameOfferList:
          type: array
          items:
            $ref: "#/definitions/GameOffer"
  GeneratedKey:
      type: object
      properties:
        id:
          type: integer
          format: int64
        generatedKey:
          type: string
        user:
          $ref: "#/definitions/ResponseUser"
        game:
          $ref: "#/definitions/Game"
        gameOrder:
          $ref: "#/definitions/GameOrder"
...
  Rating:
    type: object
    properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        game:
          $ref: "#/definitions/Game"
        rating:
          type: number
          format: double